arrange input type components and clean input file
