+++ example/demo.js
@@ -26,17 +26,13 @@ create({
     RadioIndicator: {},
     SuccessIndicator: {},
     CountIndicator: {},
-    Checkbox: {},
-    CheckboxWithTitleParagraph: {},
-    CheckboxWithLabel: {},
-    CheckboxWithImgWithLabel: {},
-    CheckBoxWithSet: {},
+    CheckBox: {},
     Radio: {},
-    RadioWithTitleParagraph: {},
     Toggle: {},
-    ToggleWithTitleParagraph: {},
     TitleParagraph: {},
-    TitleParagraphWithButton: {},
+    CheckBoxTitleParagraph: {},
+    RadioTitleParagraph: {},
+    ToggleTitleParagraph: {},
     Label: {},
     NotificationIndicator: {},
     CodeField: {},
