--- uikit/domql/Check/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-export * from './Checkbox'
-export * from './Radio'
-export * from './Toggle'
