+++ uikit/domql/index.js
@@ -19,7 +19,6 @@ export * from '@symbo.ls/field'
 export * from '@symbo.ls/search'
 export * from '@symbo.ls/button'
 export * from '@symbo.ls/progress'
-export * from '@symbo.ls/check'
 export * from '@symbo.ls/range'
 export * from '@symbo.ls/slider'
