+++ uikit/domql/package.json
@@ -15,7 +15,6 @@
   "dependencies": {
     "@symbo.ls/atoms": "latest",
     "@symbo.ls/indicator": "latest",
-    "@symbo.ls/check": "latest",
     "@symbo.ls/radiobutton": "latest",
     "@symbo.ls/toggleswitch": "latest",
     "@symbo.ls/titleparagraph": "latest",
